requirements.txt
opencv-python==4.4.0.42
numpy==1.18.5
Pillow==7.0.0
pywinauto==0.6.8